Part 5
Play with default pipes and structural directives. (make strings uppercase and create a list of elements that displays elements with an even index (0, 2, 4…). Try with a list of strings first. If this is working fine you can also try with a list of dates (Date type) (https://angular.io/api?type=pipe )

For experienced users: create a custom pipe and a custom directive (they should do basic stuff, like changing the color or adding a hyphen (-) at the beginning and the end of a string).