Part 5
Create a dummy component and route to it with a parameter. In this component take parameter from route and show it in html.
Style the list of notes to look like in the desired application
Add validations for Add Note form (using validation attributes - see example here)
For experienced students: Change template-driven form to a reactive form in Add Note component. Use validation (see example here)