Homework 7
Add data-binding on the add-note component.
Implement add note functionality using the NoteService.
Create a service for filters/categories with a method which will return the available list of filters/categories and use it in filter component.
Add a select box element on “add-note” page and populate this select box with the list of filters/categories from service. This allows you to set a category on a note. (see select box example here).